requirements.txt
flask==3.0.2
selectorlib==0.16.0
requests==2.31.0
python-dateutil==2.9.0.post0
fake-useragent==1.5.1
firebase-admin==6.5.0
scikit-learn==1.4.2
google-generativeai==0.5.2